Add debugging and rename the args to be clearer
This refactor adds detection of a DEBUG env variable and only
prints the debug if it exists and is set to something.  Also, moving
some code around so it makes more sense, and renaming the args so that
it is obvious which is an arg and which is a calculated field.