Switch to Polysemy and parse the siteconfig and header
This switches to using Polysemy for effects.  The header of the file is
also optionally parsed, and the siteconfig is now parsed using Polysemy
for effects (including throw).  Note that the header is still a raw
header and so needs further work.  Also we need to add tests, which will
be intimidating, I suspect!