Add exploration of using Polysemy and Conduit together
Essentially, an exploration of how to combine Conduit (which is the
streaming library) and Polysemy (for free effects).  This take on it,
lets Conduit just use actions on the `Sem r` monad, which feels the
least coupled way to do it.  Unfortunately, there is a lot of
boilerplate around the Members on each function.