Switch Header.hs to ByteString and add in the File searching
This switches the Header.hs to use ByteString rather than Text, which
seems to be a better fit.  The yaml loader wants a ByteString and the
other libraries (Pandoc and Ginger) are also happy with ByteString.

Also adds the The File effect (which stats files, loads content and
writes files), and FilesConduit which (although horribly named) scans
a directory and finds the relevant extension files into a list.  And it
uses a combination of Conduit and Polysemy.