Add initial explorator ginger/polysemy integration
This is an experiment in trying to use Ginger (Jinja2 type rendering)
and Polysemy together.  The objective is to get to a point where the
Polysemy Context effect will have all the information necessary to
provide context to rendering a file.

This experiment is just the start of finding out how it might fit
together.