Add ExplorePandoc.hs -- walking the AST to convert wikilinks
So, this is about converting [[link|desc]] and [[link]] to
Pandoc Link constructions by walking the AST.  This is more robust that
regexing the source md due to codeblocks and the like.  This will
replace the regex done for the v2html command (which will be split
out into a separate repo).

Next, is to re-write the links using slugs for the final layout, and
place the converting vimwiki links code into a library.