Remove the vm2html parts of the program
The vm2html generator is now a separate repository.  There may yet be a
shared library module as vw2html needs to grow the ability to recognise
the source page header to be able to strip it out when generating a
file, if a user wishes to use both sitegen and vw2html with the same
wiki.