Additional info on output (404 pages, template searching)
Additional notes on output, attempting to include details about 404
pages and a change to how templates are searched for.  Inspired by Hugo,
there is a now _default directory that is used rather than searching up
the directory.  i.e. it now looks in three places; either at the
directory where the page will be (mirrored in the templates) or in the
_default directory equivalent, or in the root of the _default directory.