Add VirtualIndexPageContext to Header.hs
The VirtualIndexPageContext record is to have a context that corresponds
to a virtual page (the index) that has not physical source page.  This
is so it can be added to the list of pages to render.  Classes (or
something) will be used to paper over the issue that there is not
body/content to actually delivery to the template.