Finally a working example!
So this commit contains a (very basic) working example of the site
generation process.  It:

 - scans the input directory for <extension> files.
 - Adds in index files
 - iterates through those files, and for each file.
   - Looks up the template
   - generates a (dynamic) ginger context
   - renders the template + context to logging output

But it's far from ready.  The error handling is all over the place, it
doesn't write to output files and there is no handling of static pages.
But it's not far off from being usable.  It's also missing a lot of
tests.

However, I've proved the use of Polysemy and Ginger mtl can be threaded
together (in a way) and that it will be able to publish a site.  So I'm
all set to continue :)