Update docs and README
This just brings the docs and README up to date so that it matches (or
at least notes where it no longer applies) the current state of the
code. In particular it includes what still needs to be done on the
README and (hopefully) sets the expectation of what the system will do.