Refactor Lib.SourceClass out of existance
Essentially, I couldn't really work out how to get it to do what I
wanted it to do.  i.e. be Polymorphic over the existential type so that
I could store SourceContext as a list, but access common variables.  It
would be possible, but it added so much complexity that I just decided
to put the relevant code into Lib.Pandoc (for the Sem r based code) and
Lib.Header for the static lookups.