Add writing output files
So this commit adds the functionality of writing files to the output.
It records all of the directories and files that it has written, creates
intermediate directories and then writes the output file.  There's still
loads of logging going on, but the next step is to copy static files
across.