Add in actual templates to check template resolution
This adds templates that do actual resolution using the site.  This
includes blocks and includes.  This resulted in a rewrite of the
template resolution code, in particular the SourceContext that only
needs to resolve to a name, but the IncludeResolver (for Ginger) needing
to do all the work about finding the template and then checking it
exists.