WIP - Get templates starting to work with variables
This patch is a WIP where the variables in templates are starting to
come together with the Contexts.  Obviously, it's throwing up lots of
issues, so this commit sort-of works, but in no way is a functional site
generator.  However, it's a good place to commit, as the next function
is pagination and that changes how pages are generated and needs a
dynamic update of the 'pages-to-render' list.  i.e. something new in
SiteGenState.