Implement a modifiable queue for the render list
This change puts the render list into a queue sorted by the route.  This
can be altered during the rendering cycle.  The key change is to enable
'plugins' (which don't really exist) to add pages to the rendering list
when a feature is accessed.  e.g. pagination of summary / titles in an
index page may require additional index pages to be created, and
index-2.html for the 2nd page, etc.  This can be done by the pagination
function when it is called to find out how many pages should be on the
index page.