Add basic pager functionality
This *may* change, but the paginate function works for a collection of
pages.  Unfortunately, there is no pagesSelect function to provide a set
of things to the paginate function.  That'll be the next step.  This
also doens't include a template example on which to build the pager to
get multiple pages.