Add non-working getlocale() function
It's not working as it requires a new effect 'Effects.Locale' that will
provide the getLocale function in the IO monad.  However, that would
require a change to every function constraint set that runs in the
RunSem r monad (which is a lot of functions).  Thus, it's time to
upgrade to the next version of polysemy (and compiler) and then change
to bundles to be more explicit about which function use which bundles of
effects.  And that's not appropriate for this commit.

Still, it's useful to have this non-working version of the function to
silence the warning about a missing function.