Update to lts-16.3 // GHC 8.8.3 // Polysemy 1.3.0.0
This updates the project to use lts-16.3 and GHC 8.8.3.  The main driver
is to be able to use Polysemy 1.3.0 which includes Bundle to group a
bunch of Effects together.

As part of this, Pandoc has been upgraded and uses Text by default,
instead of String.  This needed a few updates to the Pandoc utils and
tests.