Bundle up the Ginger Sem monad Sem r constraints into a Type
This patchset bundles all of the constraints for the Sem r monad into a
single type to simplify the type declarations for the various Ginger
context functions and to make it easier to add/remove effects.  This
ensures that the monad supplied to ginger is consistent across all
functions.  It's not really needed, but it is nice as it means only
modifying a constraint in one place and then every function in the
Context gets the effect if it needs it.