Add Effect.Locale to fully implement Ginger getlocale
Ginger expects a 'getlocale' to be implemented by the using function if
the 'dateformat' (aka date) function is used in a template.  This
patchset adds the Effect.Locale with a `getLocale` function that does
the deed, and finalises the Context function 'getlocale' to use it, and
thread the Locale effect through to the context.