Add Effect.Cache for polymorphic caching of things
This effect caches any Haskell type into a Hash or (if encode and
decode functions are provide) into a ByteString file store of temporary
files.  It will replace the ByteStringStore as it's effectively the same
thing and we don't have to marshall to and from ByteStrings when caching
in memory.