Remove Effect.ByteStringStore and optimise template parsing
The ByteStringStore is effectively void as it can be replaced by a
Cache ByteString and do exactly the same thing.  Horray for
generalisation!

Also, a slight re-write on the Ginger template parsing as there is now a
Cache (Template SourcePos) which can cache parsed templates.  The idea
is that the program only needs to parse each exact template once and
then they can be re-used as in-memory items.  We'll change to parsing
them all in upfront, perhaps, as the errors will be easier to control.
However, that means we might end up doing unnecessary parsing if the
files never get written.