Add a Logging Effect that builds on colog
This adds a basic logging effect that allows for severity and optional
module logging.  It will also make it possible to capture callstack
logging for (error/warning perhaps) in the future or as a app log rather
than user log.

Going to switch to the Output effect for printing messages, and use the
logging effect to log errors, debug, and other stuff.