Add category handling and page generation
This commit adds generation of the category page and sub pages if the
categories template asks for a category page to be generated.  It's a
bit rough and ready, but seems to work.

Also added a `params` to the VirtualPageContext, but it's unused as
present.  It was going to be for the category, but there was a better
way of doing it (detecting the categories path).

TODO: take the magic string "/categories" and put it in a type in
something like `Types.Constants` so that they can be changed more
easily.