Ensure that pages rendered are added to the rendered list
Ensure that the routes that have been rendered are added to the render
list.  This is to prevent dynamically generated routes from being cyclic
as they can be avoided.  Maybe have to add code to the RenderUtils
module to prevent and error out if the same route is going to be
rendered twice.