Major refactor of SPC and VPC to SourceMetadata
The SourcePageContext and VirtualPageContext are refactored into a
single SourceMetadata record type.  The reason they were seperate was an
early design decision where it was thought that there would be multiple
different *SourceContexts for each type of page.  It turns out that
isn't true, and the system can work very well with just a single type.

Also, renaming it to SourceMetadata explains what it is: data about the
source.  Anyway, this simplifies the software as there are only a few
places that really care about the difference between a concrete page
(one deriving from a markdown file) and a virtual page (one created by
the system).