Add permalink support to the header for pages
The `permalink` item in the header of a page allows absolute placement
of a page's route. This overrides the route completely in terms of links
and where the page is put.  However, the route still determines where
the page 'lives' from the perspective of sibling pages (for lists) and
for determining the termplate for the page.