Change Category to Categories in the page header
But keep the old category tag.  I made a decision early on to only have
a single category per page.  Turns out that that was an error as my
archive pages (in one site) have multiple categories per page.  Ho Hum.
So I added a categories tag for the header, but allow for backwards
compatibility in using a single category and adding a single tag.