Add templates override dir and skylighting support
Pandoc supports skylighting which is highlighting using just HTML and
CSS - no JS required.  This just requires a function to get the CSS into
the template, which is called styleToCSS() in the template.

Also allow template overrides dirctories to be specified so that the
theme can essentially be the overall layout and CSS styles, and then use
template override directories to override those themes.  This allows for
overriding any part of the theme as needed.  Useful for multiple sites.