Add Atom feed generation (and other fixes)
Quite a big commit here.  Mostly about adding atom.feed generation to
the system, which in turn added the ability for 'default' templates to
be built into the executable.  That required a few changes.  It also
required an Effect.Time module to be added so that the time can be got
at, and this had to be threaded through to the Ginger Context.  It also
exposed a bug in the plain text processing which had to be fixed in the
pandoc processing.

Lens have been added to the build, but aren't being used anywhere yet.
I have a feeling I'm going to back out the Polysemy code (particularly
from the Ginger processing), but maybe the whole project.  And this will
end up with more nested records on an RWS Monad stack.  Thus, Lens will
come in useful.

With this commit, it's finally ready to replicate/replace my hugo site
generation for one of my sites.