Initial stages of adding a dev-server (http)
This is to allow serving of pages whilst editing them.  Initially it's
just to do re-rendering when the source file changes, but eventually it
would use hot-reloading.  This allows the edit-view-edit cycle to be
quicker.

One issue is that I think this might need to be folded into the
(possible) successor of this.